Revert file system changes
